TrackerTableBody and TableConfig broken out. Added state and functions to handle adding and editing form dialogs
